requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.2
sqlalchemy==2.0.36
aiosqlite==0.20.0
openai==1.54.3
python-dotenv==1.0.1
pydantic-settings==2.6.1
httpx==0.27.2
websockets==12.0
redis==5.0.1
chromadb==0.4.22

